+++ functions/_PDCLIB/_PDCLIB_naive_ptod.c
@@ -11,10 +11,11 @@
 #include <ctype.h>
 #include <string.h>
 
-long double _PDCLIB_naive_ptod( const char * s, char ** endptr )
+// TODO changed long double to doubles
+double _PDCLIB_naive_ptod( const char * s, char ** endptr )
 {
     /* This is nowhere good enough, just a quick approximation */
-    long double rc = 0.0;
+    double rc = 0.0;
     const char * x;
 
     while ( ( x = (const char *)memchr( _PDCLIB_digits, tolower( (unsigned char)*s ), 16 ) ) != NULL )
@@ -24,8 +25,8 @@ long double _PDCLIB_naive_ptod( const char * s, char ** endptr )
 
     if ( *s == '.' )
     {
-        long double fraction = 0.0;
-        long double scale = 1.0;
+        double fraction = 0.0;
+        double scale = 1.0;
         ++s;
 
         while ( ( x = (const char *)memchr( _PDCLIB_digits, tolower( (unsigned char)*s ), 16 ) ) != NULL )
@@ -40,8 +41,8 @@ long double _PDCLIB_naive_ptod( const char * s, char ** endptr )
     if ( tolower( (unsigned char)*s ) == 'p' )
     {
         char sign = '+';
-        long double exp = 0.0;
-        long double scale = 1.0;
+        double exp = 0.0;
+        double scale = 1.0;
         ++s;
 
         if ( *s == '-' )
